Hugo'ify the 404 page
GitHub Pages allows for custom 404 pages, so create one using the original
as a template. Also, by using the Hugo-defined default site styling, the
unstyled copy of the table of contents can be removed in favor of rendering
the 404 content with the common header, footer, and sidebar.

Only custom 404 pages are supported, but not 500 ones: For details, see
https://docs.github.com/en/pages/getting-started-with-github-pages/creating-a-custom-404-page-for-your-github-pages-site

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>