# This is a combination of 9 commits. # The 1st commit message will be skipped:
# ci: update the books via a GitHub workflow
#
# With this commit, the SHAs corresponding to the various repositories
# containing the ProGit Book and its translations are stored in a
# sparse-checkout'able directory.
#
# This information is then used by a scheduled workflow to determine what
# needs to be updated (if anything) and then performing that task.
#
# Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

# The commit message #2 will be skipped:

# fixup! ci: update the books via a GitHub workflow

# The commit message #3 will be skipped:

# fixup! ci: update the books via a GitHub workflow

# This is the commit message #4:

# amend! ci: update the books via a GitHub workflow

ci: update the books via a GitHub workflow

With this commit, the SHAs corresponding to the various repositories
containing the ProGit Book and its translations are stored in a
sparse-checkout'able directory.

This information is then used by a scheduled workflow to determine what
needs to be updated (if anything) and then performing that task.

When GitHub workflows push new changes, they cannot trigger other GitHub
workflows (to avoid infinite loops). Therefore, this new GitHub workflow
not only synchronizes the books, but also builds the site and deploys
it.

Note: The code to build the site and to deploy it is provided in a
custom Action, to make it reusable. It will come in handy over the next
commits, where other GitHub workflows are added that likewise need
to synchronize changes that desire a site rebuild & deployment.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

# The commit message #5 will be skipped:

# fixup! fixup! ci: update the books via a GitHub workflow

# The commit message #6 will be skipped:

# fixup! fixup! ci: update the books via a GitHub workflow

# The commit message #7 will be skipped:

# fixup! ci: update the books via a GitHub workflow

# The commit message #8 will be skipped:

# fixup! ci: update the books via a GitHub workflow

# The commit message #9 will be skipped:

# fixup! ci: update the books via a GitHub workflow