Add a workflow to update the Git version and manual pages
Whenever a new Git version is released, we will want git-scm.com to be
updated automatically.

To this end, add a scheduled workflow that does precisely that. In case
someone gets too impatient to wait for the daily check, there's also a
`workflow_dispatch` trigger.

Note: It would be tempting to use a partial clone of git.git for this,
but that would cause many, many round-trips (each time a blob is
requested), especially when force-rebuilding all versions of all manual
pages (which will be added in a later commit). So let's not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>