Revert "book2.rake: drop progit2-ms translation"
Contrary to the claim of that commit, that repository _does_ exist.

Besides, it is referenced in quite a few places, e.g. in
`layouts/partials/translations.html`, the file formerly known as
`app/views/books/_translations.html.erb`. Therefore we would either
remove those references (that were forgotten in the commit that
removed the Malaysian translation), or reinstate the entry in
`@all_books`.

This reverts commit 0f82f6f5eb00d32e2347fae0c6cadc89d4f2d627.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>