book: work around lagging translations' image paths
At some stage in 2017, the progit2 layout was changed so that all images
live in a central location. The `saveImage` method was written with that
layout in mind.

However, there is at least one translation that has not switched to the
new layout yet, and its images are in a location relative to the
respective chapters' source files that reference them.

Let's make sure that we handle those layouts correctly, too, then.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>