docs: add fall-back redirects for unrendered pages
There are quite a couple of pages we do not render, e.g. all of the
`howto/` ones.

The existing Rails App can get away with trying to redirect all unknown
pages in `docs/` to the git/git repository via a generic wildcard
pattern.

In the static website, we do not have that luxury. But we can at least
add corresponding redirects for missing pages that are linked to.

While at it, work around a couple incorrect constructs such as
`linkgit:../<...>` and `linkgit:technical/<...>`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>